% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_resolution.R
\name{filename_resolution}
\alias{filename_resolution}
\title{A function to support fl_read() function}
\usage{
filename_resolution(dfcol)
}
\arguments{
\item{dfcol}{dataframe column to be operated upon}
}
\description{
This function allows you to read all the data files that reside within a nested folder structure.
}
\keyword{csv,}
\keyword{xls}
\keyword{xlsx,}
