% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_read.R
\name{fl_read}
\alias{fl_read}
\title{A Data read-in Function}
\usage{
fl_read(pth, i = 1)
}
\arguments{
\item{pth}{The path of the parent folder for the nested data files to be read from}

\item{i}{A number (optional)}
}
\description{
Allows you to read all the data files that reside within a nested folder structure.\cr
Provide the starting folder path and the system will read all relevant files within this nested folder.\cr \cr
1.	Essentially, this package was designed to read files of the format (csv/xlsx/xls), so any other files that reside within the folder structure, eg. PDF, txt, etc; will be ignored.\cr
2.	The naming convention of the read files will be:\cr
  a.	For CSV:	 ProjData_<FolderName1>_<FolderName2>_<FolderName3>._filename\cr
  b.	For XLSX/XLS:	ProjData_<FolderName1>_<FolderName2>_<FolderName3>._filename_SheetName1\cr
                      ProjData_<FolderName1>_<FolderName2>_<FolderName3>._filename_SheetName2\cr
3.	A report of files which could not be read (maybe due to corrupt data) will be generated for reference in the end.\cr
}
\examples{
\dontrun{
fl_read("Path of parent folder")

# Will read all the data files (in csv/excel format) present in this nested folder structure.

# To view the error log of unread files, use command:
err_list
}

}
\keyword{csv,}
\keyword{xls}
\keyword{xlsx,}
