% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_pkgs.R
\name{ld_pkgs}
\alias{ld_pkgs}
\title{A function to support fl_read to install required packages}
\usage{
ld_pkgs()
}
\description{
This function allows you to read all the data files that reside within a nested folder structure.
}
\keyword{csv,}
\keyword{xls}
\keyword{xlsx,}
